requirements.txt
streamlit 
python-dotenv
python-pptx 
googletrans==3.1.0a0 
aiohttp #This is for telegram

# tqdm
# comtypes 
# pywin32
# gtts
# groq
